classe per eliminazione workflow
