+++ frontend/src/hooks/useDeleteWorkflow.test.tsx
@@ -2,7 +2,7 @@ import { expect, test, describe, beforeEach, vi } from "vitest";
 import { render, screen, waitFor, act } from "@testing-library/react";
 import { AuthContextType, authProviderRender, providerPropsInit } from "../context/MockedAuthProvider";
 import { useDeleteWorkflow } from "./useDeleteWorkflow";
-import { deleteWorkflowByName } from "../services/deleteWorkflowAPI";
+import { DeleteWorkflowService } from "../services/deleteWorkflowService";
 import "@testing-library/jest-dom";
 
 vi.mock("../services/allWorkflowAPI", () => ({
@@ -37,7 +37,7 @@ describe("useDeleteWorkflow hook", () => {
     };
 
     test("Deletes a workflow successfully when user is authenticated", async () => {
-        vi.mocked(deleteWorkflowByName).mockResolvedValue({ name: "Deleted Successfully" });
+        vi.spyOn(DeleteWorkflowService.prototype, "deleteWorkflowByName").mockResolvedValue({ name: "Deleted Successfully" });
 
         authProviderRender(<TestComponent workflowName="Test Workflow" />, providerProps);
 
@@ -52,7 +52,8 @@ describe("useDeleteWorkflow hook", () => {
     });
 
     test("Handles errors when API call fails", async () => {
-        vi.mocked(deleteWorkflowByName).mockRejectedValue(new Error("API Error"));
+        vi.spyOn(DeleteWorkflowService.prototype, "deleteWorkflowByName").mockRejectedValue(new Error("API Error"));
+        
 
         authProviderRender(<TestComponent workflowName="Failing Workflow" />, providerProps);
