+++ frontend/src/services/deleteWorkflowService.test.ts
@@ -1,5 +1,5 @@
 import { expect, test, describe, vi, beforeEach } from "vitest";
-import { deleteWorkflowByName } from "./deleteWorkflowAPI";
+import { DeleteWorkflowService } from "./deleteWorkflowService";
 import { API_BASE_URL } from "./constants";
 
 describe("deleteWorkflowByName API", () => {
@@ -9,17 +9,19 @@ describe("deleteWorkflowByName API", () => {
         fetchSpy.mockReset();
     });
 
+    const name = "testWorkflow";
+    const accessToken = "testToken";
+    const service = new DeleteWorkflowService(accessToken);
+
     test("Should return the deleted workflow when successful", async () => {
-        const name = "testWorkflow";
-        const accessToken = "testToken";
-        const mockResponse = { name, accessToken };
+        const mockResponse = { name };
         
         fetchSpy.mockResolvedValue({
             status: 200,
             json: () => Promise.resolve(mockResponse)
         } as Response);
         
-        await expect(deleteWorkflowByName(name, accessToken)).resolves.toEqual(mockResponse);
+        await expect(service.deleteWorkflowByName(name)).resolves.toEqual(mockResponse);
         expect(fetchSpy).toBeCalledTimes(1);
         expect(fetchSpy).toBeCalledWith(`${API_BASE_URL}/workflow/delete/${name}`, {
             method: "DELETE",
@@ -31,35 +33,26 @@ describe("deleteWorkflowByName API", () => {
     });
 
     test("Should throw an error with message 'Generic error' if status 404 is received", async () => {
-        const name = "testWorkflow";
-        const accessToken = "testToken";
-        
         fetchSpy.mockResolvedValue({
             status: 400,
             json: () => Promise.resolve("Generic error")
         } as Response);
         
-        await expect(deleteWorkflowByName(name, accessToken)).rejects.toThrowError("Generic error");
+        await expect(service.deleteWorkflowByName(name)).rejects.toThrowError("Generic error");
     });
 
     test("Should throw an error with message 'Server error' if status 500 is received", async () => {
-        const name = "testWorkflow";
-        const accessToken = "testToken";
-        
         fetchSpy.mockResolvedValue({
             status: 500,
             json: () => Promise.resolve("Server error")
         } as Response);
         
-        await expect(deleteWorkflowByName(name, accessToken)).rejects.toThrowError("Server error");
+        await expect(service.deleteWorkflowByName(name)).rejects.toThrowError("Server error");
     });
 
     test("Should throw an error with message 'Generic error' if fetch fails", async () => {
-        const name = "testWorkflow";
-        const accessToken = "testToken";
-        
         fetchSpy.mockRejectedValue(new Error("Generic error"));
         
-        await expect(deleteWorkflowByName(name, accessToken)).rejects.toThrowError("Generic error");
+        await expect(service.deleteWorkflowByName(name)).rejects.toThrowError("Generic error");
     });
 });
